+++ test/Par/ParYaml.jl
@@ -1,7 +1,5 @@
 # Par/ParYaml
 
-ext = Serde.Ext.YAML()
-
 @testset verbose = true "ParYaml" begin
     @testset "Case №1: Simple YAML" begin
         exp_str = """
@@ -80,7 +78,7 @@ ext = Serde.Ext.YAML()
 
     @testset "Case №4: Exceptions tests" begin
         exp_str = "vector: [3,,4]"
-        @test_throws ext.ParYaml.YamlSyntaxError Serde.parse_yaml(exp_str)
+        @test_throws Serde.YAML.YamlSyntaxError Serde.parse_yaml(exp_str)
     end
 
     @testset "Case №5: Different Dict type" begin
@@ -89,6 +87,6 @@ ext = Serde.Ext.YAML()
         bar: true
         """
         exp_obj = IdDict("foo" => 163, "bar" => true)
-        @test ext.ParYaml.parse_yaml(exp_str; dict_type = IdDict) == exp_obj
+        @test Serde.YAML.parse_yaml(exp_str; dict_type = IdDict) == exp_obj
     end
 end
